requirements.txt
python-fasthtml
functions-framework==3.*